Almost-server detection and reporting (#249)
This PR provides warnings for any server failing validation and outputs
the malformed or incorrect parts of the config. It also adds a new
optional env variable `error_on_almost_servers` for a ValueError to be
raised.

Currently in the gym spinup process, any server(s) failing validation
results in the server getting silently dropped and raising a generic
AssertionError.

---------

Signed-off-by: Frankie Siino <fsiino@nvidia.com>